refactor(Delivery.Web): move context types out of Context namespace
