+++ src/delivery/Sandbox.Delivery.Web/Features/PageMetas/PageMetaStandardizer.cs
@@ -3,7 +3,7 @@ using System.Linq;
 using Ardalis.GuardClauses;
 using Sandbox.Core.Domain.Intrastructure.Operations.Queries;
 using Sandbox.Delivery.Core.Features.MarketingTagsContents;
-using Sandbox.Delivery.Web.Infrastructure.Contexts;
+using Sandbox.Delivery.Web.Infrastructure.Core;
 
 namespace Sandbox.Delivery.Web.Features.PageMetas
 {
