+++ src/delivery/Sandbox.Delivery.Web/Infrastructure/Core/DocumentContext.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Sandbox.Delivery.Web.Infrastructure.Contexts
+namespace Sandbox.Delivery.Web.Infrastructure.Core
 {
     public interface IDocumentContext
     {
