+++ src/delivery/Sandbox.Delivery.Web/Infrastructure/Core/DocumentQueryContext.cs
@@ -1,7 +1,8 @@
 ﻿using Ardalis.GuardClauses;
 using Sandbox.Data.Kentico.Infrastructure.Queries;
+using Sandbox.Delivery.Web.Infrastructure.PageBuilders;
 
-namespace Sandbox.Delivery.Web.Infrastructure.Contexts
+namespace Sandbox.Delivery.Web.Infrastructure.Core
 {
     public class DocumentQueryContext : IDocumentQueryContext
     {
