+++ src/delivery/Sandbox.Delivery.Web/Infrastructure/Core/HttpContextBaseAccessor.cs
@@ -1,6 +1,6 @@
 ﻿using System.Web;
 
-namespace Sandbox.Delivery.Web.Infrastructure.Contexts
+namespace Sandbox.Delivery.Web.Infrastructure.Core
 {
     public interface IHttpContextBaseAccessor
     {
