+++ src/delivery/Sandbox.Delivery.Web/Infrastructure/Core/PreviewContext.cs
@@ -2,14 +2,14 @@
 using Kentico.Content.Web.Mvc;
 using Kentico.Web.Mvc;
 
-namespace Sandbox.Delivery.Web.Infrastructure.PageBuilders
+namespace Sandbox.Delivery.Web.Infrastructure.Core
 {
     public interface IPreviewContext
     {
         bool IsPreviewEnabled { get; }
     }
 
-    public class KenticoPreviewContext : IPreviewContext
+    public class PreviewContext : IPreviewContext
     {
         public bool IsPreviewEnabled => HttpContext.Current.Kentico().Preview().Enabled;
     }
